requirements.txt
streamlit
openai
requests
pillow
python-dotenv
langdetect
google-generativeai
twilio